Django/django6일차/oz-backend-django1/pyproject.toml
[project]
name = "oz-backend-django"
version = "0.1.0"
description = "python django project"
authors = [
    {name = "ParkKangbin",email = "6883997@naver.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
